split Awards and Review components, removed some div styling, added 2rem padding to awards
